+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.0-dev
   * Bug Fixes
     * register will not return the current state for the key. Instead, the registrant will queued a message.
+    * added update_in/3 for modifying the value at a scope in place.
 
 ## v0.2.1
   * Bug Fixes